Make mod spawn eggs compatible with other mods
This commit changes the LazySpawnEggItem to extends the vanilla
SpawnEggItem class. Also, the mod spawn eggs are added to the
SpawnEggItem.EGGS map via the FMLCommonSetupEvent. This means that some
duplicated code can be removed from the LazySpawnEggItem class, and more
importantly, this mods items are now in the vanilla EGGS map, which
means that other mods can find them.